文章分类删除功能

+++ public/assets/js/category.js
@@ -45,7 +45,7 @@ $('#categoryBox').on('submit', '#editCategory', function () {
         type: 'put',
         url: `/categories/${id}`,
         data: formData,
-        success: function(response) {
+        success: function (response) {
             location.reload();
         }
     });
@@ -53,52 +53,54 @@ $('#categoryBox').on('submit', '#editCategory', function () {
 })
 
 // 删除分类
-$('#categoryList').on('click', '.delete', function() {
-    let id = $(this).attr('data-id');
-    $.ajax({
-        type: 'delete', 
-        url: `/categories/${id}`,
-        success: function () {
-            location.reload();
-        }
-    })
+$('#categoryList').on('click', '.delete', function () {
+    if (confirm('确定要删除吗？')) {
+        let id = $(this).attr('data-id');
+        $.ajax({
+            type: 'delete',
+            url: `/categories/${id}`,
+            success: function () {
+                location.reload();
+            }
+        })
+    }
 })
 
 // 批量删除用户
 let selectAll = $("#selectAll");
 let deleteMany = $("#deleteMany");
 // 全选按钮控制单选按钮
-selectAll.on('change', function() {
+selectAll.on('change', function () {
     let status = $(this).prop('checked');
-    if(status) {
+    if (status) {
         deleteMany.show();
-    }else{
-        deleteMany.hide();  
+    } else {
+        deleteMany.hide();
     }
     $('#categoryList').find('input').prop('checked', status);
 })
 // 单个按钮控制全选按钮
-$('#categoryList').on('change', '.categoryStatus', function() {
+$('#categoryList').on('change', '.categoryStatus', function () {
     let inputs = $('#categoryList').find('input');
-    if(inputs.length == inputs.filter(':checked').length) {
+    if (inputs.length == inputs.filter(':checked').length) {
         selectAll.prop('checked', true);
-    }else{
+    } else {
         selectAll.prop('checked', false);
     }
-    if(inputs.filter(':checked').length > 0) {
+    if (inputs.filter(':checked').length > 0) {
         deleteMany.show();
-    }else{
+    } else {
         deleteMany.hide();
     }
 })
-deleteMany.on('click', function() {
+deleteMany.on('click', function () {
     let checkedCategory = $('#categoryList').find('input:checked');
     let ids = [];
-    checkedCategory.each((i,ele) => {
+    checkedCategory.each((i, ele) => {
         ids.push($(ele).attr('data-id'));
     })
     console.log(ids);
-    if(confirm('确定要批量删除吗')) {
+    if (confirm('确定要批量删除吗')) {
         $.ajax({
             type: 'delete',
             url: `/categories/${ids.join('-')}`,
